Move stripComments and handle non ".json"
